When dropping privileges, also set current groups to the user default groups
Without this change, the process running the shell does not have the
same group access as the user specified in the config, and therefore
cannot give that access to child shells. For example, since the the
process running sshd does not belong to `docker`, it cannot give
`docker` group access to child shells.