Add support for remote commands
Run multiple poll loops. One for remote commands, another for remote
sessions. Currently with the same polling interval configuration.

An MQTT poll notification will poll for both sessions and commands.

Added two configuration options:

- `device.commands_dir` - This needs to point to a writable
directory. RAC will keep state about running commands in this
directory so it can properly notify the server about the state of
execution.

- `device.commands_timeout` - The default timeout for running
commands. If the command do not finish in this interval, it will be
interrupted and a failure will be reported.

Currently the following commands are supported:

- reboot - The command `sudo systemctl reboot` will be executed. The
reboot will be scheduled for 30 seconds later, and the successful
scheduling of the command will cause the command to be considered
successfully executed and that status will be reported back to RAS.

- restart-service - The service given in argument will be restarted
via `sudo systemctl restart <service>`.

Notice that both commands will use sudo, since RAC is assumed to be
running with an unprivileged user. The sudoers file needs to be
appropriately provisioned for this execution to be successful.

Both the arguments and the commands will be verified against the
metadata available in an uptane validated `remote-sessions.json`
metadata.

See the `commands.rs` file for the failure modes covered by this
implementation when running remote commands.